add autosave functionality

+++ frontend/src/components/TodoListForm.jsx
@@ -7,44 +7,39 @@ import DeleteIcon from '@mui/icons-material/Delete'
 import AddIcon from '@mui/icons-material/Add'
 import { saveListTodos, deleteTodo } from '../modules/todos'
 import { calculateRemainingTime } from '../utils/time'
+import _ from 'lodash'
 
 export const TodoListForm = ({ todoList }) => {
+  const { id: listId } = todoList
   const [todos, setTodos] = useState([...todoList.todos]);
 
   const handleAddTodo = () => {
     setTodos([...todos, { id: '', description: '', dueBy: new Date() }]);
   };
 
-  const handleInputChange = (index, event) => {
-    const updatedTodos = [...todos];
-    updatedTodos[index].description = event.target.value;
-    setTodos(updatedTodos);
-  };
+  const debouncedSave = _.debounce(async (listId, todos) => {
+    await saveListTodos({ listId, todos });
+  }, 200);
 
-  const handleDueByChange = (index, date) => {
+  const handleTodoChange = (index, field, value) => {
     const updatedTodos = [...todos];
-    updatedTodos[index].dueBy = date;
+    updatedTodos[index][field] = value;
     setTodos(updatedTodos);
+    debouncedSave(listId, updatedTodos);
   };
 
   const handleDeleteTodo = async (id) => {
-    await deleteTodo(id);
-    setTodos(todos.filter((todo) => todo.id !== id));
-  };
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    await saveListTodos({ listId: todoList.id, todos });
+    if (window.confirm("Are you sure you want to delete this todo?")) {
+      await deleteTodo(id);
+      setTodos(todos.filter((todo) => todo.id !== id));
+    }
   };
 
   return (
     <Card sx={{ margin: '0 1rem' }}>
       <CardContent>
         <Typography component='h2'>{todoList.title}</Typography>
-        <form
-          onSubmit={handleSubmit}
-          style={{ display: 'flex', flexDirection: 'column', flexGrow: 1 }}
-        >
+        <form style={{ display: 'flex', flexDirection: 'column', flexGrow: 1 }}>
           {todos.map((todo, index) => {
             const { id, description, dueBy } = todo
             const dueByDate = new Date(dueBy);
@@ -57,13 +52,13 @@ export const TodoListForm = ({ todoList }) => {
                   sx={{ flexGrow: 1, marginTop: '1rem' }}
                   label='What to do?'
                   value={description}
-                  onChange={(event) => handleInputChange(index, event)}
+                  onChange={(event) => handleTodoChange(index, 'description', event)}
                 />
                 <LocalizationProvider dateAdapter={AdapterDateFns}>
                   <DatePicker
                     label='Due by'
                     value={dueByDate}
-                    onChange={(date) => handleDueByChange(index, date)}
+                    onChange={(date) => handleTodoChange(index, 'dueBy', date)}
                     slots={{ textField: (params) => <TextField {...params} /> }}
                   />
                 </LocalizationProvider>
@@ -89,9 +84,6 @@ export const TodoListForm = ({ todoList }) => {
               onClick={() => handleAddTodo()}
             >
               Add Todo <AddIcon />
-            </Button>
-            <Button type='submit' variant='contained' color='primary'>
-              Save
             </Button>
           </CardActions>
         </form>
